Add nidcpower module (#387)
* Add initial nidcpower metadata

This was extracted by NI’s internal tools, exported as
nidcpower-nimi-python/export/0.1/0.1.0d1

* Add initial nidcpower metadata

This was extracted by NI’s internal tools, exported as
nidcpower-nimi-python/export/0.1/0.1.0d1

* Modify for NI-DCPower (came from NI-DMM)

* Make replace_func_python_name() more tolerant to typos in metadata documentation

Instead of dying, it warns and keeps going. This was due in order to
workaround typos in NI-DCPower CVI FP file from which our metadata is
extracted.

* Changes to include docs for nidcpower

* Add nidcpower to README

* Add ViInt64

NI-DCPower needs it

* Update CHANGELOG.md

* Fix function name

* Remove unused script

It’s stale and a maintenance burden. I assume nobody uses it.

* Add nidcpower to build scripts and makefiles.

* Placeholder example files

* Manual fixes to generated functions.py

This is necessary until internal NI bug 669900 is fixed.

* Placeholder system tests

* Update generated

* Merge from master

* Retrofit for generator changes

* Update generated

* Fix merge mistake

* add pass so interpreter is happy if you run

* add missing --append flag

* Add AttributeViInt64 class.

Tests not written yet.

* Flake 8

* Update generated

* Fixing flake8 failures on examples

* Fixing flake8 failures on system tests

* Adding functions addon

* Removing tBoolean enum

* Updating tBoolean to be None and removing hidden attributes

* Updating generated