Update HistoryRAMCycleInformation.__repr__ to use `__module__` and `__qualname__`
Best practice for implementing `__repr__` for custom types is to use
`__module__` and `__qualname__`.

By including `__module__` in `__repr__`, `eval(repr(HistoryRAMCycleInformation()))`
can be called by importing nidigital as `import nidigital`; which is
consistent with recommended import style of nimi-python.

If this change is accepted, `__repr__` for private and internal types
(including nifake) will be changed in a separate PR. `__repr__` for public
types in other modules will be done in 2.0 Source Breakers milestone.