Add clear links to usage sections in documentation (#1720)
* Add links to GitHub locations of examples in usage documentation

Including the link to GitHub clarifies that these examples are not
installed with the rest of the module. This reduces possible confusion
on the user's part, as they may look for examples in the installed
package and decide that the documentation is untrustworthy.

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Render documentation

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Shorten sentence to avoid redundancy

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Render documentation

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Remove unused file

This file appears to be completely unused. In fact, running
`git log -G'[^_]usage\.inc'` suggests that it has *never* been used.

Signed-off-by: Mike Petersen <mike.petersen@ni.com>