Note configuration must be passed to manual invocation of pytest
Without providing the configuration to pytest, it will perform an
[automatic search][1] to locate a configuration, and erroneously pick up
the top-level tox.ini rather than the correct tox-system_tests.ini
generated for each driver. That causes many `PytestUnknownMarkWarning`s
to be signaled when running the test, as noted in #1655.

This change only touches CONTRIBUTING.md, and only the instructions for
invoking pytest. Other changes will be required to get tox working in a
similar manner, as it appears that it does not inform pytest of the
configuration file to use, even if tox itself uses the correct (and
indeed identical) one.

[1]: <https://docs.pytest.org/en/6.2.x/customize.html#finding-the-rootdir>

Signed-off-by: Mike Petersen <mike.petersen@ni.com>