Add configuration file argument to system test execution
When tox executes the system tests, it dispatches to coverage, which
itself dispatches to pytest. As noted in the previous change, pytest
does not by default receive guidance about which configuration file to
use, causing `PytestUnknownMarkWarning`s to be signaled while running
the system tests. This change adds the argument to coverage (passed on
to pytest) to set the configuration file to use, so that pytest does not
rely upon its search (which would not travel to a cousin directory, nor
determine that this file name would be acceptable).

Signed-off-by: Mike Petersen <mike.petersen@ni.com>