Use correct pytest configuration file (#1731)
* Note configuration must be passed to manual invocation of pytest

Without providing the configuration to pytest, it will perform an
[automatic search][1] to locate a configuration, and erroneously pick up
the top-level tox.ini rather than the correct tox-system_tests.ini
generated for each driver. That causes many `PytestUnknownMarkWarning`s
to be signaled when running the test, as noted in #1655.

This change only touches CONTRIBUTING.md, and only the instructions for
invoking pytest. Other changes will be required to get tox working in a
similar manner, as it appears that it does not inform pytest of the
configuration file to use, even if tox itself uses the correct (and
indeed identical) one.

[1]: <https://docs.pytest.org/en/6.2.x/customize.html#finding-the-rootdir>

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Add configuration file argument to system test execution

When tox executes the system tests, it dispatches to coverage, which
itself dispatches to pytest. As noted in the previous change, pytest
does not by default receive guidance about which configuration file to
use, causing `PytestUnknownMarkWarning`s to be signaled while running
the system tests. This change adds the argument to coverage (passed on
to pytest) to set the configuration file to use, so that pytest does not
rely upon its search (which would not travel to a cousin directory, nor
determine that this file name would be acceptable).

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Generate per-driver configuration files for system tests

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Fix typo in tox invocation of system tests

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Remove superfluous space

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Use forward slashes in tox invocation example

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Rename NI-TClk system test script to have correct prefix

The Python script to run the NI-TClk system tests began with a `tests_`
prefix, while the automatic detection of such files searches for a
`test_` prefix in its [standard test discovery][1].

This was noted because the previous changes caused the system tests to
run using their actual per-driver configuration files, and the top-level
tox.ini that had previously been used has this line in its `[pytest]`
section:

    python_files = *.py

which caused *all* files to be picked up and considered to be test
scripts.

[1]: <https://docs.pytest.org/en/6.2.x/goodpractices.html#test-discovery>

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Remove redundant parent directory traversals for system tests

Since the system tests are executed with the current directory already
being the generated per-driver directory, it is unnecessary to travel
upwards just to return to where we started. This makes these
configuration files a bit easier to read.

Signed-off-by: Mike Petersen <mike.petersen@ni.com>

* Copy pytest flags from top-level tox.ini to tox-system_tests.ini.mako

Now that the system tests are no longer using the top-level tox.ini file
for their configuration, it is desirable to have some of those flags
copied over so that the test output is consistent with what it was
previously.

Signed-off-by: Mike Petersen <mike.petersen@ni.com>